Upgrade to Laravel 5.4

+++ database/migrations/2016_09_20_092248_add_user_activations.php
@@ -25,8 +25,10 @@ class AddUserActivations extends Migration
      */
     public function down()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('activation_token');
-        });
+        if (Schema::hasColumn('users', 'activation_token')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('activation_token');
+            });
+        }
     }
 }
